Add weather page and setting
Add a setting for the weather zip.  Add a weather page to load weather
based on the setting.  Update the weather inline widget to display
weather for the current location based on IP address - the widget
provider does not have a way to allow the user to specify the location.